Auth_Manager/requirements.txt
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
cryptography==36.0.2
dnspython==2.2.1
Flask==2.1.1
Flask-Cors==3.0.10
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
PyJWT==2.3.0
MarkupSafe==2.1.1
pycparser==2.21
pymongo==4.1.0
requests==2.27.1
six==1.16.0
urllib3==1.26.9
Werkzeug==2.1.1
zipp==3.8.0
